% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jossl-package.R
\docType{package}
\name{jossl-package}
\alias{jossl-package}
\title{Joint Survival Super Learner}
\description{
Joint Survival Super Learner
}
\author{
\strong{Maintainer}: Anders Munch \email{a.munch@sund.ku.dk}

Authors:
\itemize{
  \item Thomas Alexander Gerds
}

}
\keyword{internal}
