% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jossl.R
\name{jossl}
\alias{jossl}
\title{Joint Survival Super Learner}
\usage{
jossl(
  learners,
  data,
  time,
  integrate = TRUE,
  split.method = "cv5",
  collapse = TRUE,
  B = 1,
  verbose = FALSE,
  time_grid_length = 100,
  time_name = "time",
  status_name = "status",
  cause_codes = c(`1` = 1, `2` = 2, c = 0),
  vars = NULL
)
}
\arguments{
\item{learners}{library of learners}

\item{data}{data}

\item{time}{time}

\item{integrate}{logical. If \code{TRUE} then integrate}

\item{split.method}{Passed on to riskRegression::getSplitMethod}

\item{collapse}{logical. If \code{TRUE} then average}

\item{B}{Passed on to riskRegression::getSplitMethod}

\item{verbose}{logical. If \code{TRUE} then talk a lot}

\item{time_grid_length}{Time grid to approximate the integrated Brier score}

\item{time_name}{Name of time variable}

\item{status_name}{Name of status variable}

\item{cause_codes}{Labels}

\item{vars}{Additional variables}
}
\value{
Fitted object with class jossl
##' @examples
set.seed(8)
library(survival)
d <- riskRegression::sampleData(87)
lol <- list(
  cox_lasso = list("cox", x_form = ~X2+X7),
  aalen_johansen =  list("cox", x_form = ~ 1))
x = jossl(list(cause1 = lol,cause2 = lol,censor = lol),
          data = d,
          time = 36,
          integrate = TRUE,
          verbose = TRUE,
          split.method = "cv5",
          B = 1,
          time_name = "time",
          status_name = "event",
          cause_codes = c("1" = 1, "2" = 2, "c" = 0),
          vars = c("X1","X2","X3","X4","X5","X6","X7"))
}
\description{
Joint survival super learner
}
\details{
Argument names may change in the future
}
\author{
Anders Munch
}
