% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCHF.R
\name{predictCHF}
\alias{predictCHF}
\alias{predictCHF.coxph}
\alias{predictCHF.GLMnet}
\alias{predictCHF.rfsrc}
\title{Predicting cumulative hazard functions}
\usage{
predictCHF(object, newdata, ...)

\method{predictCHF}{coxph}(object, newdata, times, ...)

\method{predictCHF}{GLMnet}(object, newdata, times, ...)

\method{predictCHF}{rfsrc}(object, newdata, times, ...)
}
\arguments{
\item{object}{Object}

\item{newdata}{newdata}

\item{...}{additional arguments}

\item{times}{times}
}
\description{
Predicting cumulative hazard functions
}
