% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRisk.jossl.R
\name{predictRisk.jossl}
\alias{predictRisk.jossl}
\title{Predict method for joint survival super learner}
\usage{
\method{predictRisk}{jossl}(object, newdata, times, cause, ...)
}
\arguments{
\item{object}{Object obtained with \code{jossl}}

\item{newdata}{Newdata}

\item{times}{Times}

\item{cause}{Cause}

\item{...}{Not used}
}
\value{
matrix with one column for each value in \code{times} and
one row for each individual in \code{newdata}
}
\description{
Predict method for joint survival super learner
}
\details{
Predicts the risks of all causes and also the censoring distribution.
}
\examples{
set.seed(8)
library(survival)
d <- riskRegression::sampleData(87)
lol <- list(
  cox_lasso = list("cox", x_form = ~X2+X7),
  aalen_johansen =  list("cox", x_form = ~ 1))
x = jossl(list(cause1 = lol,cause2 = lol,censor = lol),
          data = d,
          time = 36,
          integrate = TRUE,
          verbose = TRUE,
          split.method = "cv5",
          B = 1,
          time_name = "time",
          status_name = "event",
          cause_codes = c("1" = 1, "2" = 2, "c" = 0),
          vars = c("X1","X2","X3","X4","X5","X6","X7"))
predictRisk.jossl(x,times = 3:8,newdata = d[1:3],cause = 1)
}
\seealso{
\link[riskRegression]{predictRisk}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
