% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.jossl.R
\name{print.jossl}
\alias{print.jossl}
\title{Print jossl object}
\usage{
\method{print}{jossl}(x, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{...}{passed to print}

\item{digits}{Number of digits}
}
\description{
Print method for jossl
}
